Update user guide (Tagging)
User guide has room for improvement.

Some readers may not find the current user guide tangible enough or that the logical flow is a bit clunky.

Let's shift up the command summary to the top first, then sequence the tables of commands from general usage to ones for specific uses. After that let's add in the feature for v1.4 with visual aids.