Standardise user guide
User guide has had a lot of recent changes

Standardisation of the user guide at least in terms of layout would make it more tangible for the reader.

Let's sort the commands in alphabetical order, then ensure that it at least follows a standard layout with some variation in wording as it is a draft.